remove framework-extra-bundle requirement: use argument resolver instead of paramconverter
