+++ docs/param-converter.md
@@ -1,6 +1,7 @@
-Fazland - DtoManagementBundle - Param Converter
-===============================================
-Now that we have a DTO `Resolver` it is easy to have also a ParamConverter.
+Fazland - DtoManagementBundle - Argument Resolver
+=================================================
+Now that we have a DTO `Resolver` it is easy to have also an argument resolver.  
+This bundle leverages the symfony's argument resolver mechanism to inject DTOs into controllers.
 
 That is, as the example in the beginning of these docs, you can directly inject the right implementation of the desired DTO into the controller action.
