+++ tests/Fixtures/Proxy/AppKernel.php
@@ -4,7 +4,6 @@ namespace Fazland\DtoManagementBundle\Tests\Fixtures\Proxy;
 
 use Fazland\DtoManagementBundle\DtoManagementBundle;
 use Fazland\DtoManagementBundle\Tests\Fixtures\TestKernel;
-use Sensio\Bundle\FrameworkExtraBundle\SensioFrameworkExtraBundle;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
 use Symfony\Bundle\SecurityBundle\SecurityBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
@@ -18,7 +17,6 @@ class AppKernel extends TestKernel
     {
         return [
             new FrameworkBundle(),
-            new SensioFrameworkExtraBundle(),
             new DtoManagementBundle(),
             new SecurityBundle(),
             new AppBundle(),
